% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCircos_wrapper.R
\name{BioCircos_wrapper}
\alias{BioCircos_wrapper}
\title{Convenience wrapper function for the BioCircos package}
\usage{
BioCircos_wrapper(data, genome = NULL, genome_color = NULL, gene_name,
  gene_start, gene_end, gene_chromosome, gene_strand = NULL,
  gene_category = NULL, gene_abundance = NULL,
  gene_foldchange = NULL, gene_heatmap = NULL, gene_linetrack = NULL,
  gene_labels = NULL, color_heatmap = c("#F5F5F5", "#4E962D"),
  range_abundance = NULL, range_foldchange = NULL,
  range_heatmap = NULL, range_linetrack = NULL, track_size = 0.15,
  track_gap = 0.02)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{genome}{A named list of chromosomes and their total length such as
\code{list(chromsome_1 = 1000)}. If not provided the genome information
is retrieved from gene annotation.}

\item{genome_color}{(character) (optional) color vector of same length as \code{genome}.}

\item{gene_name}{(character) column name of the data frame containing gene names.}

\item{gene_start}{(character) column name of the data frame containing start positions in bp.}

\item{gene_end}{(character) column name of the data frame containing end positions in pb.}

\item{gene_chromosome}{(character) column name of the data frame containing the chromosome ID of each gene.}

\item{gene_strand}{(character) (optional) column name of the data frame containing the strand of each gene ('+' or '-').}

\item{gene_category}{(character) (optional) column name of the data frame containing gene annotation such 
as 'Respiratory chain' or 'Unknown'.}

\item{gene_abundance}{(character) (optional) column name of the data frame containing an absolute abundance 
such as protein length or mass fraction.}

\item{gene_foldchange}{(character) (optional) column name of the data frame containing fold change values for a gene.}

\item{gene_heatmap}{(character) (optional) column name(s) of the data frame containing abundances plotted as heatmap. 
Can contain more than one element.}

\item{gene_linetrack}{(character) (optional) column name(s) of the data frame containing abundances plotted as lines.
Can contain more than one element.}

\item{gene_labels}{(character) (optional) column name with labels that will be plotted as arcs. 
Can contain NAs for labels hat should not be plotted.}

\item{color_heatmap}{(character) (optional) vector of length 2 indicating colors to use for heatmap gradient}

\item{range_abundance}{(numeric) (optional) range for plotting gene/protein abundances. Vector of length 2.}

\item{range_foldchange}{(numeric) (optional) range for plotting gene/protein fold changes. Vector of length 2.}

\item{range_heatmap}{(numeric) (optional) range for plotting gene/protein abundances. Vector of length 2.}

\item{range_linetrack}{(numeric) (optional) range for plotting gene/protein abundances. Vector of length 2.}

\item{track_size}{(numeric) (optional) size of one circular layer added to the plot, a scalar between 0 and 1.}

\item{track_gap}{(numeric) (optional) size of the gap between circular layers, a scalar between 0 and 1.}
}
\description{
This function takes as arguments a data.frame and a list of columns
}
