% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCircos.R
\name{BioCircosLinkTrack}
\alias{BioCircosLinkTrack}
\title{Create an inner track with links to be added to a BioCircos tracklist}
\usage{
BioCircosLinkTrack(trackname, gene1Chromosomes, gene1Starts, gene1Ends,
  gene2Chromosomes, gene2Starts, gene2Ends, color = "#40B9D4", labels = "",
  maxRadius = 0.4, width = "0.1em", gene1Names = "", gene2Names = "",
  ...)
}
\arguments{
\item{trackname}{The name of the new track.}

\item{gene1Chromosomes, gene1Starts, gene1Ends, gene1Names, gene2Chromosomes, gene2Starts, gene2Ends, gene2Names}{Vectors with the chromosomes, genomic coordinates of beginning and end, and names of both genes to link.
Chromosomes and positions should respect the chromosome names and lengths given in the genome parameter of
the BioCircos function.}

\item{color}{The color for the links, in hexadecimal RGB format.}

\item{labels}{A vector of character objects to label each link.}

\item{maxRadius}{Where the track should end, in proportion of the inner radius of the plot.}

\item{width}{The thickness of the links.}

\item{...}{Ignored}
}
\description{
Links are defined by beginning and ending genomic coordinates of the 2 regions to linked,
 such as the positions linked in genomic fusions.
}
