% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCircos.R
\name{BioCircosSNPTrack}
\alias{BioCircosSNPTrack}
\title{Create a track with SNPs to be added to a BioCircos tracklist}
\usage{
BioCircosSNPTrack(trackname, chromosomes, positions, values)
}
\arguments{
\item{trackname}{The name of the new track.}

\item{chromosomes}{A vector containing the chromosomes on which each SNP are found.
Values should match the chromosome names given in the genome parameter of the BioCircos function.}

\item{positions}{A vector containing the coordinates on which each SNP are found.
Values should be inferior to the chromosome lengths given in the genome parameter of the BioCircos function.}

\item{values}{A vector of numerical values associated with each SNPs, used to determine the 
radial coordinates of each point on the visualization.}

\item{tracklist}{The list of tracks of your BioCircos visualization.}
}
\description{
SNPs are defined by genomic coordinates and associated with a numerical value
}
