% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCircos.R
\name{BioCircosTracklist}
\alias{BioCircosTracklist}
\alias{+.BioCircosTracklist}
\title{Create a list of BioCircos tracks}
\usage{
BioCircosTracklist()

\method{+}{BioCircosTracklist}(x, ...)
}
\description{
This allows the use of the '+' operator on these lists
}
