% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.panel.functions.R
\name{panel.barplot}
\alias{panel.barplot}
\title{Draw barplot with error bars in lattice plots}
\usage{
panel.barplot(x, y, ewidth = 0.08,
  border = trellis.par.get()$superpose.polygon$col[1], fill = "white",
  ...)
}
\description{
Custom panel functions for lattice plots
}
