% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.panel.functions.R
\name{panel.errbars}
\alias{panel.errbars}
\title{Calculate and draw error bars in lattice plots}
\usage{
panel.errbars(x, y, ewidth = 0.08, FUN = function(x) mean(x, na.rm =
  TRUE), border = trellis.par.get()$superpose.polygon$col[1], ...)
}
\description{
Custom panel functions for lattice plots
}
