% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.panel.functions.R
\name{panel.key}
\alias{panel.key}
\title{Draw custom keys in lattice plots}
\usage{
panel.key(labels, which.panel = 1, pch = 1, cex = 0.8,
  point.cex = NULL, points = TRUE, lines = FALSE,
  rectangles = FALSE,
  col = trellis.par.get()$superpose.polygon$col[1:length(labels)],
  lwd = trellis.par.get()$superpose.line$lwd[1],
  lty = trellis.par.get()$superpose.line$lty[1], corner = c(0.1, 0.9),
  x = corner[1], y = corner[2], ...)
}
\description{
Custom panel functions for lattice plots
}
