% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.panel.functions.R
\name{panel.pvalue}
\alias{panel.pvalue}
\title{Calculate and draw p-values in lattice plots}
\usage{
panel.pvalue(x, y, std, symbol = TRUE, cex.symbol = 1.5, offset = 1,
  fixed.pos = NULL, verbose = TRUE,
  col = trellis.par.get()$superpose.polygon$col[1], ...)
}
\description{
Custom panel functions for lattice plots
}
