% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.panel.functions.R
\name{panel.quadrants}
\alias{panel.quadrants}
\title{Draw quadrants and quadrant statistics in lattice plots}
\usage{
panel.quadrants(x, y, h = NULL, v = NULL, labels = "percent",
  col = grey(0.5), margin = 0.1,
  lwd = trellis.par.get()$superpose.polygon$lwd[1], ...)
}
\description{
Custom panel functions for lattice plots
}
