% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCircos.R
\name{setCircInvisible}
\alias{setCircInvisible}
\title{Assign the supplied edge attribute values to the graph structure contained in the browser.}
\usage{
setCircInvisible(session)
}
\arguments{
\item{session}{a Shiny Server session object.}
}
\description{
Assign the supplied edge attribute values to the graph structure contained in the browser.
}
\examples{
\dontrun{
  setCircInvisible(session)
}

}
