% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCircos.R
\name{setCircNodeVisibility}
\alias{setCircNodeVisibility}
\title{Assign the supplied node visibility values to the nodes on the circle}
\usage{
setCircNodeVisibility(session, nodes, values)
}
\arguments{
\item{session}{a Shiny Server session object.}

\item{nodes}{a character vector the names of the nodes whose attributes are updated.}

\item{values}{a character, logical or numeric vector, the new values.}
}
\description{
Assign the supplied node visibility values to the nodes on the circle
}
\examples{
\dontrun{
  setNodeAttributes(session,
                    nodes=yeastGalactodeNodeIDs,
                    values=expression.vector)
}

}
