% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioCircos.R
\name{BioCircosTracklist}
\alias{BioCircosTracklist}
\alias{+.BioCircosTracklist}
\alias{-.BioCircosTracklist}
\title{Create a list of BioCircos tracks}
\usage{
BioCircosTracklist()

\method{+}{BioCircosTracklist}(x, ...)

\method{-}{BioCircosTracklist}(x, ...)
}
\arguments{
\item{x}{The tracklist on which other tracks should be added or removed.}

\item{...}{The tracks to add (as tracklists) or to remove (as track names).}
}
\description{
This allows the use of the '+' and '-' operator on these lists
}
